refactored model into engine and game folders
