+++ src/app/model/engine/faction.model.ts
