+++ src/app/model/engine/game-state-store.model.ts
