+++ src/app/model/game/element.model.ts
